Moved eveything to app folder
